refactor: move CoordinateSystem class to new coord_system package
