+++ gmatpyplus/orbit_state.py
@@ -1,152 +1,10 @@
 from __future__ import annotations
 
 import gmatpyplus as gp
-from basics import GmatObject
-from gmatpyplus import gmat
-from utils import Barycenter, CelestialBodies, CoordSystems, gmat_str_to_py_str, GroundStations, LibrationPoints, \
-    SpacecraftObjs, py_str_to_gmat_str
+from utils import CoordSystems, gmat_str_to_py_str, py_str_to_gmat_str
 
 
 class OrbitState:
-    class CoordinateSystem(GmatObject):
-        # TODO convert __init__ params to args with default values
-
-        # TODO complete - will be able to create each type of Axes, for use in CoordinateSystem
-        class Axes(GmatObject):
-            def __init__(self, axes_type: str, name: str):
-                super().__init__(axes_type, name)
-                self.Initialize()
-
-        def __init__(self, name: str, origin: str = 'Earth', axes: str = 'MJ2000Eq', primary: str = None,
-                     secondary: str = None, xaxis: str = None, yaxis: str = None, zaxis: str = None, epoch: str = None,
-                     alignment_vec_x: int = None, alignment_vec_y: int = None, alignment_vec_z: int = None,
-                     constraint_vec_x: int = None, constraint_vec_y: int = None, constraint_vec_z: int = None,
-                     constraint_ref_vec_x: int = None, constraint_ref_vec_y: int = None,
-                     constraint_ref_vec_z: int = None,
-                     constraint_coord_sys: str = None, ref_object: str = None
-                     ):
-            # TODO: remove kwargs if possible, if not document as another 2do
-            # TODO complete allowed values - see User Guide pages 335-339 (PDF pg 344-348)
-            #  and src/base/coordsystem/CoordinateSystem.cpp/CreateLocalCoordinateSystem
-            super().__init__('CoordinateSystem', name)
-            self.allowed_values = {'Axes': ['MJ2000Eq', 'MJ2000Ec', 'ICRF',
-                                            'MODEq', 'MODEc', 'TODEq', 'TODEc', 'MOEEq', 'MOEEc', 'TOEEq', 'TOEEc',
-                                            'ObjectReferenced', 'Equator', 'BodyFixed', 'BodyInertial',
-                                            'GSE', 'GSM', 'Topocentric', 'BodySpinSun'],
-                                   'CentralBody': CelestialBodies(),
-                                   'Origin': (CelestialBodies() + SpacecraftObjs() + LibrationPoints() + Barycenter()
-                                              + GroundStations()),
-                                   'AxesTypeSpecific': {
-                                       'ObjectReferenced': {
-                                           'Primary': (CelestialBodies() + SpacecraftObjs() + LibrationPoints() +
-                                                       Barycenter() + GroundStations()),
-                                           'Secondary': (CelestialBodies() + SpacecraftObjs() + LibrationPoints() +
-                                                         Barycenter() + GroundStations()),
-                                           'XAxis': ['R', 'V', 'N', '-R', '-V', '-N', None],
-                                           'YAxis': ['R', 'V', 'N', '-R', '-V', '-N', None],
-                                           'ZAxis': ['R', 'V', 'N', '-R', '-V', '-N', None],
-                                       },
-                                       'TOE': {
-                                           'Epoch': '21545'
-                                       },
-                                       'MOE': {
-                                           'Epoch': '21545'
-                                       },
-                                       'LocalAlignedConstrained': {
-                                           'AlignmentVectorX': 1,
-                                           'AlignmentVectorY': 0,
-                                           'AlignmentVectorZ': 0,
-                                           'ConstraintVectorX': 0,
-                                           'ConstraintVectorY': 0,
-                                           'ConstraintVectorZ': 1,
-                                           'ConstraintReferenceVectorX': 0,
-                                           'ConstraintReferenceVectorY': 0,
-                                           'ConstraintReferenceVectorZ': 1,
-                                           'ConstraintCoordinateSystem': 'EarthMJ2000Eq',
-                                           'ReferenceObject': (CelestialBodies() + SpacecraftObjs() +
-                                                               LibrationPoints() + Barycenter() +
-                                                               GroundStations())
-                                       }
-                                   },
-                                   }
-
-            # Parse origin argument
-            if origin not in self.allowed_values['Origin']:
-                raise AttributeError(f'Specified origin "{origin}" is not recognized. Please specify one of the '
-                                     f'following:\n\t{self.allowed_values["Origin"]}')
-            else:
-                self.origin = gmat.GetObject(origin)  # get current (default) origin
-                # attach new origin to CoordinateSystem
-                self.SetStringParameter(1, self.origin.GetName())  # 1 for ORIGIN_NAME, 2 for J2000_BODY_NAME
-                self.SetRefObject(self.origin, gmat.SPACE_POINT, self.origin.GetName())
-
-            # Parse axes argument
-            if axes not in self.allowed_values['Axes']:
-                raise AttributeError(f'Specified axes type "{axes}" is not recognized. Please specify one of the '
-                                     f'following:\n\t{self.allowed_values["Axes"]}')
-            else:
-                if axes in list(self.allowed_values['AxesTypeSpecific'].keys()):
-                    axes_specific_values = self.allowed_values['AxesTypeSpecific'][axes]
-
-                    # TODO set params/ref objs for all axes types
-                    if axes == 'ObjectReferenced':
-                        self.primary = primary
-                        self.secondary = secondary
-                        self.xaxis = xaxis
-                        self.yaxis = yaxis
-                        self.zaxis = zaxis
-
-                    elif (axes == 'TOE') or (axes == 'MOE'):
-                        self.epoch = epoch
-
-                    elif axes == 'LocalAlignedConstrained':
-                        self.alignment_vec_x = alignment_vec_x
-                        self.alignment_vec_y = alignment_vec_y
-                        self.alignment_vec_z = alignment_vec_z
-                        self.constraint_vec_x = constraint_vec_x
-                        self.constraint_vec_y = constraint_vec_y
-                        self.constraint_vec_z = constraint_vec_z
-                        self.constraint_ref_vec_x = constraint_ref_vec_x
-                        self.constraint_ref_vec_y = constraint_ref_vec_y
-                        self.constraint_ref_vec_z = constraint_ref_vec_z
-                        self.constraint_coord_sys = constraint_coord_sys
-                        self.ref_object = ref_object
-
-            self.axes: OrbitState.CoordinateSystem.Axes = OrbitState.CoordinateSystem.Axes(axes, f'{origin}_{axes}')
-            self.SetRefObject(self.axes, gmat.AXIS_SYSTEM, self.axes.name)
-
-            # gp.Initialize()
-            self.Initialize()
-
-        def __repr__(self):
-            return f'A CoordinateSystem with origin {self.origin} and axes {self.axes}'
-
-        @staticmethod
-        def Construct(name: str, central_body: str, axes: str):
-            print('In static Construct')
-            return gmat.Construct('CoordinateSystem', name, central_body, axes)
-
-        @classmethod
-        def from_sat(cls, sc: gp.Spacecraft) -> OrbitState.CoordinateSystem:
-            name = sc.gmat_obj.GetRefObjectName(gmat.COORDINATE_SYSTEM)
-            sc_cs_gmat_obj = sc.gmat_obj.GetRefObject(gmat.COORDINATE_SYSTEM, name)
-            origin = sc_cs_gmat_obj.GetField('Origin')
-            axes = sc_cs_gmat_obj.GetField('Axes')
-            coord_sys: OrbitState.CoordinateSystem = cls(name=name, origin=origin, axes=axes)
-            return coord_sys
-
-        @property
-        def name(self) -> str:
-            return self._name if self._name else self.gmat_obj.GetName()
-
-        @name.setter
-        def name(self, name):
-            self._name = name
-            self.gmat_obj.SetName(name)
-
-        def Help(self):
-            return GmatObject.Help(self.gmat_obj)
-
     def __init__(self, **kwargs):
         self.allowed_state_elements = {
             'Cartesian': {'X', 'Y', 'Z', 'VX', 'VY', 'VZ'},
